split text package into grapheme and input
