+++ internal/ui/text/grapheme.go
@@ -1,31 +1,11 @@
+// Package text provides grapheme-aware helpers shared by the TUI overlays:
+// cluster boundary navigation for text editing and key-message-to-text
+// mapping for input fields.
 package text
 
 import (
-	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/x/ansi"
 )
-
-// PrintableInput returns the printable text represented by a key message.
-// Bubble Tea reports a regular space as KeySpace, while other text uses
-// KeyRunes and may contain more than one rune when pasted or entered via IME.
-func PrintableInput(msg tea.KeyMsg) (string, bool) {
-	if msg.Type != tea.KeyRunes && msg.Type != tea.KeySpace {
-		return "", false
-	}
-	s := string(msg.Runes)
-	if msg.Type == tea.KeySpace {
-		s = " "
-	}
-	if s == "" {
-		return "", false
-	}
-	for _, r := range s {
-		if r < 32 {
-			return "", false
-		}
-	}
-	return s, true
-}
 
 // PreviousGraphemeBoundary returns the byte index before the grapheme cluster
 // at pos, or -1 when pos is already at the start.
